+++ buildSrc/src/main/kotlin/versions.kt
@@ -1,13 +1,20 @@
 object vers {
+    val kotlin = "1.4.21"
+    val nexus_staging = "0.22.0"
+
     object asoft {
-        val color = "0.0.1"
-        val theme = "0.0.1"
+        val builders = "1.3.0"
+        val color = "0.0.20"
+        val theme = "0.0.50"
+        val test = "1.1.10"
+    }
+
+    object kotlinx {
+        val coroutines = "1.4.2"
     }
 
     object wrappers {
-        val react = "17.0.0-pre.129-kotlin-1.4.10"
-        val styled = "5.2.0-pre.129-kotlin-1.4.10"
-        val css = "1.0.0-pre.129-kotlin-1.4.10"
-        val react_router_dom = "5.2.0-pre.129-kotlin-1.4.10"
+        val react = "17.0.1-pre.141-kotlin-1.4.21"
+        val styled = "5.2.0-pre.141-kotlin-1.4.21"
     }
 }