+++ theme-core/src/commonMain/kotlin/tz/co/asoft/Theme.kt
@@ -3,7 +3,7 @@ package tz.co.asoft
 /**
  * @property T is the typography of the respective platform
  */
-data class Theme<T : Any>(
+open class Theme<T : Any>(
     val name: String,
     val color: ColorPalette,
     val text: T