+++ theme-core/src/jsMain/kotlin/tz/co/asoft/AquaGreenTheme.kt
@@ -1,9 +1,0 @@
-package tz.co.asoft
-
-val AquaGreenTheme by lazy {
-    Theme(
-        name = "aQua Green",
-        color = AquaGreenPallet,
-        text = Typography()
-    )
-}