+++ theme-css/src/commonMain/kotlin/tz/co/asoft/AquaGreenTheme.kt
@@ -1,0 +1,7 @@
+package tz.co.asoft
+
+fun AquaGreenTheme(typography: Typography? = null) = Theme(
+    name = "aQua Green [${typography?.name ?: "default"}]",
+    color = AquaGreenPallet,
+    text = typography ?: Typography(),
+)