+++ theme-css/src/commonMain/kotlin/tz/co/asoft/DarkGrayTheme.kt
@@ -1,0 +1,7 @@
+package tz.co.asoft
+
+fun DarkGrayTheme(typography: Typography? = null) = Theme(
+    name = "Dark Gray [${typography?.name ?: "default"}]",
+    color = DarkGrayPallet,
+    text = typography ?: Typography(),
+)