+++ theme-react/src/main/kotlin/tz/co/asoft/ThemeConsumer.kt
@@ -1,0 +1,8 @@
+package tz.co.asoft
+
+import react.RBuilder
+import react.createContext
+
+val ThemeContext by lazy { createContext(currentTheme.value) }
+
+inline fun RBuilder.ThemeConsumer(noinline handler: RBuilder.(ReactTheme) -> Unit) = ThemeContext.Consumer(handler)